// src/module/qna/interface/qna.interface.ts
import { QnaCategoryEnum, YNEnum } from '../../../common/constant/enum';

export interface QnaInfo {
  qnaIdx: number;
  userIdx: number;
  userName: string;
  userCell: string;
  category: QnaCategoryEnum;
  text: string;
  replySuccessYN: YNEnum;
  replyText: string;
  replyAdmin: string;
  createdAt: Date;
}